Move WebsocketClient and WebsocketClientDelegate to a file
